+++ src/XMLDocument.cpp
@@ -180,7 +180,7 @@ XMLElement *XMLElementList::Item(int nIndex)
 {
 	CAutoLock elementsLock(&m_elementsLock);
 
-	if ((nIndex < 0) || (nIndex >= m_elements.size()))
+	if ((nIndex < 0) || (nIndex >= (int)m_elements.size()))
 		return NULL;
 	return m_elements.at(nIndex);
 }
